requirements.txt
# Data Pipeline
pandas>=2.0.0
numpy>=1.22.0
yfinance>=0.2.18
polygon-api-client>=1.12.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.20.0

# Models
scikit-learn>=1.1.0
statsmodels>=0.13.0
pymc>=5.0.0
tensorflow>=2.12.0
keras>=2.12.0
torch>=2.0.0

# Reinforcement Learning
gymnasium>=0.26.0
stable-baselines3>=2.0.0
optuna>=3.1.0

# API
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Frontend
streamlit>=1.22.0
plotly>=5.14.0
dash>=2.9.0

# Utils
tqdm>=4.64.0
matplotlib>=3.7.0
seaborn>=0.12.0
ta>=0.10.0 